+++ skill-db/frontend/src/components/EditComponents/ConfirmButton.vue
@@ -1,0 +1,13 @@
+<template>
+
+</template>
+
+<script>
+export default {
+  name: "ConfirmButton"
+}
+</script>
+
+<style scoped>
+
+</style>
